fix : 권한 요청 알림 이미지 후 바로 권한 요청하도록 변경 (정책)
