+++ RunWithMe/presentation/src/main/java/com/side/runwithme/view/MainActivity.kt
@@ -29,7 +29,6 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main){
     private lateinit var navController : NavController
     
     override fun init() {
-        requestPermission()
 
         runningCheck()
 
@@ -54,85 +53,6 @@ class MainActivity : BaseActivity<ActivityMainBinding>(R.layout.activity_main){
             }
         }
     }
-
-    @RequiresApi(Build.VERSION_CODES.Q)
-    private fun permissionDialog(){
-        var builder = AlertDialog.Builder(this)
-        builder.setTitle("백그라운드 위치 권한을 위해 항상 허용으로 설정해주세요.")
-
-        var listener = DialogInterface.OnClickListener { _, type ->
-            when (type) {
-                DialogInterface.BUTTON_POSITIVE -> backgroundPermission()
-            }
-        }
-        builder.setPositiveButton("네", listener)
-        builder.setNegativeButton("아니오", null)
-
-        builder.show()
-    }
-
-    @RequiresApi(Build.VERSION_CODES.Q)
-    private fun backgroundPermission(){
-        ActivityCompat.requestPermissions(
-            this,
-            arrayOf(
-                Manifest.permission.ACCESS_BACKGROUND_LOCATION,
-            ), 2)
-    }
-
-    private fun requestPermission(){
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-            requestPermissionUpTo33()
-        }else {
-            requestPermissionUnder33()
-        }
-    }
-
-    @RequiresApi(Build.VERSION_CODES.TIRAMISU)
-    private fun requestPermissionUpTo33(){
-        TedPermission.create()
-            .setPermissionListener(object : PermissionListener {
-
-                override fun onPermissionGranted() {
-                    if (checkSelfPermission(Manifest.permission.ACCESS_BACKGROUND_LOCATION)
-                        == PackageManager.PERMISSION_DENIED) {
-                        permissionDialog()
-                    }
-                }
-                override fun onPermissionDenied(deniedPermissions: List<String>) {
-                    showToast("권한을 허가해주세요.")
-                }
-            })
-            .setDeniedMessage("앱 사용을 위해 권한을 허용으로 설정해주세요. [설정] > [앱 및 알림] > [고급] > [앱 권한]")
-            .setPermissions(
-                Manifest.permission.ACCESS_COARSE_LOCATION,
-                Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.POST_NOTIFICATIONS
-            )
-            .check()
-    }
-    private fun requestPermissionUnder33(){
-        TedPermission.create()
-            .setPermissionListener(object : PermissionListener {
-                @RequiresApi(Build.VERSION_CODES.Q)
-                override fun onPermissionGranted() {
-                    if (checkSelfPermission(Manifest.permission.ACCESS_BACKGROUND_LOCATION)
-                        == PackageManager.PERMISSION_DENIED) {
-                        permissionDialog()
-                    }
-                }
-                override fun onPermissionDenied(deniedPermissions: List<String>) {
-                    showToast("권한을 허가해주세요.")
-                }
-            })
-            .setDeniedMessage("앱 사용을 위해 권한을 허용으로 설정해주세요. [설정] > [앱 및 알림] > [고급] > [앱 권한]")
-            .setPermissions(
-                Manifest.permission.ACCESS_COARSE_LOCATION,
-                Manifest.permission.ACCESS_FINE_LOCATION
-            )
-            .check()
-    }
-
 
     private fun initNavigation() {
         val navHostFragment = supportFragmentManager.findFragmentById(R.id.fragment_container_view) as NavHostFragment
